+++ packages/sdk/test/OperationRequestBuilder.test.ts
@@ -14,6 +14,7 @@ import {
   monkey,
   plutocracy,
   getDummyHex,
+  DUMMY_CONTRACT_ADDR,
 } from "./utils";
 
 describe("OperationRequestBuilder", () => {
@@ -28,7 +29,7 @@ describe("OperationRequestBuilder", () => {
       refundAssets: [shitcoin],
       actions: [
         {
-          contractAddress: "0x1234",
+          contractAddress: DUMMY_CONTRACT_ADDR,
           encodedFunction: getDummyHex(0),
         },
       ],
@@ -38,7 +39,7 @@ describe("OperationRequestBuilder", () => {
 
     const builder = new OperationRequestBuilder();
     const opRequest = builder
-      .action("0x1234", getDummyHex(0))
+      .action(DUMMY_CONTRACT_ADDR, getDummyHex(0))
       .unwrap(shitcoin, 3n)
       .refundAsset(shitcoin)
       .chainId(1n)
@@ -67,7 +68,7 @@ describe("OperationRequestBuilder", () => {
       refundAssets: [shitcoin],
       actions: [
         {
-          contractAddress: "0x1234",
+          contractAddress: DUMMY_CONTRACT_ADDR,
           encodedFunction: getDummyHex(0),
         },
       ],
@@ -77,7 +78,7 @@ describe("OperationRequestBuilder", () => {
 
     const builder = new OperationRequestBuilder();
     const opRequest = builder
-      .action("0x1234", getDummyHex(0))
+      .action(DUMMY_CONTRACT_ADDR, getDummyHex(0))
       .unwrap(shitcoin, 3n)
       .refundAsset(shitcoin)
       .confidentialPayment(shitcoin, 1n, receiver)
@@ -103,7 +104,7 @@ describe("OperationRequestBuilder", () => {
       refundAssets: [shitcoin],
       actions: [
         {
-          contractAddress: "0x1234",
+          contractAddress: DUMMY_CONTRACT_ADDR,
           encodedFunction: getDummyHex(0),
         },
       ],
@@ -117,7 +118,7 @@ describe("OperationRequestBuilder", () => {
 
     const builder = new OperationRequestBuilder();
     const opRequest = builder
-      .action("0x1234", getDummyHex(0))
+      .action(DUMMY_CONTRACT_ADDR, getDummyHex(0))
       .unwrap(shitcoin, 3n)
       .refundAsset(shitcoin)
       .refundAddr(refundAddr)
@@ -194,7 +195,7 @@ describe("OperationRequestBuilder", () => {
       .map((signer) => signer.canonicalAddress());
 
     const actions = range(2).map((i) => ({
-      contractAddress: "0x1234",
+      contractAddress: DUMMY_CONTRACT_ADDR,
       encodedFunction: getDummyHex(i),
     }));
     const expected: OperationRequest = {
@@ -241,8 +242,8 @@ describe("OperationRequestBuilder", () => {
 
     const builder = new OperationRequestBuilder();
     const opRequest = builder
-      .action("0x1234", getDummyHex(0))
-      .action("0x1234", getDummyHex(1))
+      .action(DUMMY_CONTRACT_ADDR, getDummyHex(0))
+      .action(DUMMY_CONTRACT_ADDR, getDummyHex(1))
       .unwrap(shitcoin, 3n)
       .unwrap(ponzi, 69n)
       .unwrap(stablescam, 420n)
@@ -267,6 +268,51 @@ describe("OperationRequestBuilder", () => {
     opRequest.joinSplitRequests.sort((a, b) =>
       a.asset.assetAddr.localeCompare(b.asset.assetAddr)
     );
+
+    expect(opRequest).to.eql(expected);
+  });
+
+  it("combines requests of same asset when no conf payments", () => {
+    const sk = generateRandomSpendingKey();
+    const signer = new NocturneSigner(sk);
+    const refundAddr = signer.generateRandomStealthAddress();
+
+    const actions = range(2).map((i) => ({
+      contractAddress: DUMMY_CONTRACT_ADDR,
+      encodedFunction: getDummyHex(i),
+    }));
+    const expected: OperationRequest = {
+      joinSplitRequests: [
+        {
+          asset: shitcoin,
+          unwrapValue: 300n,
+        },
+        {
+          asset: ponzi,
+          unwrapValue: 300n,
+        },
+      ],
+      refundAssets: [],
+      refundAddr: refundAddr,
+      actions,
+      chainId: 1n,
+      deadline: 2n,
+    };
+
+    const builder = new OperationRequestBuilder();
+    const opRequest = builder
+      .action(DUMMY_CONTRACT_ADDR, getDummyHex(0))
+      .action(DUMMY_CONTRACT_ADDR, getDummyHex(1))
+      .unwrap(shitcoin, 100n)
+      .unwrap(ponzi, 100n)
+      .unwrap(shitcoin, 100n)
+      .unwrap(ponzi, 100n)
+      .unwrap(shitcoin, 100n)
+      .unwrap(ponzi, 100n)
+      .refundAddr(refundAddr)
+      .chainId(1n)
+      .deadline(2n)
+      .build();
 
     expect(opRequest).to.eql(expected);
   });