+++ packages/sdk/test/utils.ts
@@ -1,4 +1,4 @@
-import { getDefaultProvider, utils } from "ethers";
+import { ethers, getDefaultProvider, utils } from "ethers";
 import {
   Asset,
   AssetTrait,
@@ -14,36 +14,50 @@ import {
 } from "../src";
 import { Handler, Handler__factory } from "@nocturne-xyz/contracts";
 
+export const DUMMY_CONTRACT_ADDR = ethers.utils.getAddress(
+  "0x67f8f9a5d4290325506b119980660624dc7d3ba9"
+);
+
 export const shitcoin: Asset = {
   assetType: AssetType.ERC20,
-  assetAddr: "0x123",
+  assetAddr: ethers.utils.getAddress(
+    "0xddbd1e80090943632ed47b1632cb36e7ca28abc2"
+  ),
   id: 0n,
 };
 export const encodedShitcoin = AssetTrait.encode(shitcoin);
 
 export const ponzi: Asset = {
   assetType: AssetType.ERC20,
-  assetAddr: "0x456",
+  assetAddr: ethers.utils.getAddress(
+    "0x6798639591530fbbafd12c2826422b58bd2c5219"
+  ),
   id: 0n,
 };
 export const encodedPonzi = AssetTrait.encode(ponzi);
 
 export const stablescam: Asset = {
   assetType: AssetType.ERC20,
-  assetAddr: "0x789",
+  assetAddr: ethers.utils.getAddress(
+    "0x6375394335f34848b850114b66a49d6f47f2cda8"
+  ),
   id: 0n,
 };
 export const encodedStablescam = AssetTrait.encode(stablescam);
 
 export const monkey: Asset = {
   assetType: AssetType.ERC721,
-  assetAddr: "0xabc",
+  assetAddr: ethers.utils.getAddress(
+    "0x42fad9d9197846ef2bfb1418e953ca449798e022"
+  ),
   id: 1n,
 };
 
 export const plutocracy: Asset = {
   assetType: AssetType.ERC1155,
-  assetAddr: "0xdef",
+  assetAddr: ethers.utils.getAddress(
+    "0x0e1136cc3a2147ca178d265ae336602217988f48"
+  ),
   id: 1n,
 };
 export const encodedPlutocracy = AssetTrait.encode(plutocracy);